# server/docs/swagger.yaml
basePath: /
definitions:
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.Patient:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name_en:
        type: string
      first_name_th:
        type: string
      gender:
        type: string
      id:
        type: integer
      last_name_en:
        type: string
      last_name_th:
        type: string
      middle_name_en:
        type: string
      middle_name_th:
        type: string
      national_id:
        type: string
      passport_id:
        type: string
      patient_hn:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Hospital Management APIs
  version: "1.0"
paths:
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with username and password
      parameters:
      - description: Login Credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successful login
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login user
  /api/patient/:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all patient records stored in the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Patient'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all patients
      tags:
      - patient
    post:
      consumes:
      - application/json
      description: Create a new patient
      parameters:
      - description: Patient data
        in: body
        name: patient
        required: true
        schema:
          $ref: '#/definitions/models.Patient'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Patient'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new patient
      tags:
      - patients
  /api/patient/{id}:
    get:
      consumes:
      - application/json
      description: Search patient by national ID or passport ID
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Patient'
        "404":
          description: Patient not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search patient by national ID or passport ID
      tags:
      - patients
schemes:
- http
- https
swagger: "2.0"